Moves PaginatedList out of Bitlink file. Documents Bitlink methods
